pyproject.toml
[project]
name = "aql-agent"
version = "0.1.0"
description = "Lightweight AI agent for Raspberry Pi 5"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "mcp>=1.0.0",
    "slack-bolt>=1.18.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
whatsapp = ["pywa[fastapi]>=2.0.0", "fastapi>=0.110.0", "uvicorn>=0.27.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aql_agent"]
